libcompat: Fix comparison of integers of different sign in obstack_blank()
We are comparing a size_t against the difference between two pointers.
Cast the latter into size_t to fix the warning.

Warned-by: clang

+++ lib/compat/obstack.h
@@ -277,7 +277,7 @@ extern int obstack_exit_failure;
 __extension__								\
 ({ struct obstack *__o = (OBSTACK);					\
    size_t __len = (length);						\
-   if (__o->chunk_limit - __o->next_free < __len)			\
+   if ((size_t) (__o->chunk_limit - __o->next_free) < __len)		\
      _obstack_newchunk (__o, __len);					\
    (void) 0; })
